backend/requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
djangorestframework-gis>=1.0
django-cors-headers>=4.0.0
channels>=4.0.0
channels-redis>=4.1.0
celery>=5.3.0
redis>=4.6.0
psycopg2-binary>=2.9.0
Pillow>=10.0.0
numpy>=1.24.0
face-recognition>=1.3.0
librosa>=0.10.0
transformers>=4.30.0
torch>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
daphne>=4.0.0